pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "levante-bench"
version = "0.1.0"
description = "LEVANTE VLM benchmark: compare VLMs to children's behavioral data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26,<3",
    "pandas>=2.2",
    "torch>=2.2",
    "Pillow>=10",
    "requests>=2.31",
    "omegaconf>=2.3",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]
# Pinned versions for reproducible installs: see requirements.txt

[project.optional-dependencies]
dev = ["pytest", "hypothesis", "ruff", "black", "isort"]
transformers = ["transformers"]
vlm = ["transformers>=5.3", "torchvision>=0.26", "num2words>=0.5"]

[project.scripts]
levante-bench = "levante_bench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
